# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuula API
  version: 0.1.0
  description: |
    Uganda Fact-Guard (Zuula). Two surfaces share this document:

    - **Core API** (`/api/v1/*`) — used by the Zuula web app. Cookie session (or bearer JWT for
      non-browser core clients). Roles: `public`, `journalist`, `expert`, `admin`, gated per
      operation.
    - **Partner API** (`/v1/*`, served from `api.zuula.ug` in production) — `Authorization: Bearer
      <key>` with an API key, 100 requests/hour. This surface is documented for external partners
      on `/developers` and this contract matches that page exactly: same base URL shape, same
      four endpoints, same field names, same `Authorization: Bearer` header (not `X-API-Key`),
      same page-based pagination shape. See "Deviations from the P2 brief" in the P2 contract
      summary sent to Noah for the two places this contract intentionally differs from the
      brief's defaults in order to match the live `/developers` page and the working frontend
      code.

    All timestamps are ISO-8601 UTC. Requirement IDs (`FR-AUTH-01`, `§9.2`, …) refer to the CIT
    *System Development Plan & Functional Requirements Specification v1.0*.
  contact:
    name: Noah Prinzy
    url: https://github.com/Noah-Prinzy/zuula
servers:
  - url: https://zuula.ug
    description: Web app origin. Core API paths already include /api/v1.
  - url: https://api.zuula.ug
...
security:
  - sessionAuth: []
paths:
...
  /api/v1/me/submissions:
    get:
      operationId: listMySubmissions
      tags: [submissions]
      summary: My submission history (FR-SUBMIT-07, Activity page).
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ActivitySubmission" }

        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a claim, link or media file for checking (FR-SUBMIT-01…05).
      description: |
        Anonymous or signed-in. Requires a Cloudflare Turnstile token when signed out
        (FR-AUTH-07). Accepts an `Idempotency-Key` header so a retried request doesn't create
        a duplicate submission.
      security: [{}, { sessionAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
        - name: Accept-Language
          in: header
          schema: { $ref: "#/components/schemas/LocaleCode" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/SubmissionInput" }
          application/json:
            schema: { $ref: "#/components/schemas/SubmissionInput" }
      responses:
        "202":
          description: Queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionAccepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMedia" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/submissions/{trackingId}:
    get:
      operationId: getSubmissionStatus
      tags: [submissions]
      summary: Submission status and, once complete, the report (FR-SUBMIT-06, FR-DETECT-07).
      security: []
      parameters:
        - { $ref: "#/components/parameters/TrackingId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionStatusResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/submissions/{trackingId}/events:
    get:
      operationId: streamSubmissionEvents
      tags: [submissions]
      summary: Server-sent events for live pipeline progress (FR-DETECT-07).
      security: []
      parameters:
        - { $ref: "#/components/parameters/TrackingId" }
      responses:
        "200":
          description: "`text/event-stream` of `SubmissionStepEvent` frames, ending with a final `SubmissionStatusResponse` event named `done`."
          content:
            text/event-stream:
              schema: { type: string }

  # ---------------------------------------------------------------------------------------
  # Fact-checks / Library / Search (FR-SEARCH-01…05, FR-DETECT, FR-EXPLAIN)
  # ---------------------------------------------------------------------------------------
...
  /v1/checks:
    servers: [{ url: https://api.zuula.ug }]
    post:
      operationId: partnerSubmitCheck
      tags: [partner]
      summary: Submit content for verification.
      security: [{ partnerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerCheckInput" }
          multipart/form-data:
            schema: { $ref: "#/components/schemas/PartnerCheckInput" }
      responses:
        "202":
          description: Queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmissionAccepted" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMedia" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/checks/{trackingId}:
    servers: [{ url: https://api.zuula.ug }]
    get:
      operationId: partnerGetCheckStatus
      tags: [partner]
      summary: Poll a check's status; `result` is present once completed.
      security: [{ partnerApiKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/TrackingId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerCheckStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }

        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /webhooks/whatsapp:
    get:
      operationId: verifyWhatsAppWebhook
      tags: [webhooks]
      summary: Meta's webhook subscription handshake.
      security: []
      parameters:
        - { name: hub.mode, in: query, schema: { type: string } }
        - { name: hub.verify_token, in: query, schema: { type: string } }
        - { name: hub.challenge, in: query, schema: { type: string } }
      responses:
        "200":
          description: Verified — echoes hub.challenge back.
          content: { text/plain: { schema: { type: string } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: receiveWhatsAppMessage
      tags: [webhooks]
      summary: Inbound WhatsApp messages create submissions.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: Acknowledged. }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: zuula_session
      description: "Core API. An opaque, server-side session token, set as this cookie on sign-in. Non-browser core clients may send the same token as `Authorization: Bearer <token>` instead (ADR 0002 §5)."
    partnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: "zl_live_*"
      description: |
        Partner API. Matches `/developers` exactly: `Authorization: Bearer zl_live_...`
        (deliberately not the `X-API-Key` header the P2 brief first proposed — see the
        deviations note in `info.description`).
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: perPage
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    TrackingId:
      name: trackingId
      in: path
      required: true
      schema: { type: string, pattern: "^ZL-[A-Z2-9]{4}-[A-Z2-9]{2}$" }
      example: ZL-7K3P-Q9
...
  responses:
    BadRequest:
      description: Bad request.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Unauthorized:
      description: Missing or invalid credentials.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    NotFound:
      description: Not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
    PayloadTooLarge:
      description: File exceeds 50 MB.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    UnsupportedMedia:
      description: Unsupported file type.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    UnprocessableEntity:
      description: Validation failed.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    RateLimited:
      description: Rate limit exceeded (100 requests/hour per key).
      headers:
        X-RateLimit-Limit: { schema: { type: integer }, description: "Requests allowed per window." }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer }, description: Unix timestamp. }
        Retry-After: { schema: { type: integer }, description: Seconds. }
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
...
    Verdict:
      type: string
      enum: [authentic, likely-false, "false", ai-generated, unverifiable]
      description: FR-DETECT-01.
    ContentType:
      type: string
      enum: [text, url, image, audio, video]
    ClaimAssessment:
      type: string
      enum: ["false", misleading, unsupported, out-of-context, supported]
    CitationStance:
      type: string
      enum: [supports, contradicts, context]
    RaterRole:
      type: string
      enum: [public, journalist, expert]
...
    LocaleCode:
      type: string
      enum: [en, lg, ach, nyn, teo]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - file_too_large
                - unsupported_media
                - invalid_content
                - rate_limited
                - server_error
            message: { type: string }
            retryAfter: { type: integer, description: Seconds. Present for rate_limited. }
      example:
        error:
          code: rate_limited
          message: This key has used its 100 requests for this hour.
          retryAfter: 1260
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        perPage: { type: integer }
        total: { type: integer }

    # ---- Fact-check report ----
    Claim:
      type: object
      required: [id, start, end, assessment, reason, citationIds]
      properties:
        id: { type: string }
        start: { type: integer, description: Character offset into submittedText. }
        end: { type: integer, description: Exclusive. }
        assessment: { $ref: "#/components/schemas/ClaimAssessment" }
        reason: { type: string }
        citationIds: { type: array, items: { type: string } }
    Citation:
      type: object
      required: [id, sourceName, title, url, publishedAt, stance, trusted]
      properties:
        id: { type: string }
        sourceName: { type: string }
        title: { type: string }
        url: { type: string, format: uri }
        publishedAt: { type: string, format: date }
        stance: { $ref: "#/components/schemas/CitationStance" }
        trusted: { type: boolean }
        excerpt: { type: string }
    AISignal:
      type: object
      required: [id, label, description, score, threshold, method]
      properties:
        id: { type: string }
        label: { type: string }
        description: { type: string }
        score: { type: number, minimum: 0, maximum: 1 }
        threshold: { type: number, minimum: 0, maximum: 1 }
        method: { type: string }
    ExpertAnnotation:
      type: object
      required: [id, author, role, createdAt, body]
      properties:
        id: { type: string }
        author: { type: string }
        role: { type: string }
        createdAt: { type: string, format: date-time }
        body: { type: string }
    HumanReview:
      type: object
      required: [outcome, reviewer, reviewedAt, justification]
      properties:
        outcome: { type: string, enum: [confirmed, overridden] }
        reviewer: { type: string }
        reviewedAt: { type: string, format: date-time }
        previousVerdict: { $ref: "#/components/schemas/Verdict" }
        justification: { type: string }
    RatingCounts:
      type: object
      required: [public, journalist, expert]
      properties:
        public: { type: integer }
        journalist: { type: integer }
        expert: { type: integer }
    RatingComment:
      type: object
      required: [id, author, role, vote, body, createdAt]
      properties:
        id: { type: string }
        author: { type: string }
        role: { $ref: "#/components/schemas/RaterRole" }
        vote: { type: string, enum: [accurate, inaccurate] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    CommunityScore:
      type: object
      description: Computed server-side (§9.1 weights, §9.2 thresholds). Not stored client-side.
      required: [ccs, weightedAccurate, weightedInaccurate, accurateCount, inaccurateCount, total, status]
      properties:
        ccs: { type: [number, "null"], minimum: 0, maximum: 100 }
        weightedAccurate: { type: number }
        weightedInaccurate: { type: number }
        accurateCount: { type: integer }
        inaccurateCount: { type: integer }
        total: { type: integer }
        status: { type: string, enum: [verified, standard, questioned, escalated, suspended] }
    CommunityRating:
      type: object
      required: [accurate, inaccurate, comments, score]
      properties:
        accurate: { $ref: "#/components/schemas/RatingCounts" }
        inaccurate: { $ref: "#/components/schemas/RatingCounts" }
        comments:
          type: array
          items: { $ref: "#/components/schemas/RatingComment" }
        score: { $ref: "#/components/schemas/CommunityScore" }
    FactCheckReport:
      description: Full report — core API only (includes submittedText, annotations and comments).
      type: object
      required:
        [
          id, trackingId, title, contentType, language, submittedText, verdict, confidence,
          summary, whatIsFalse, whatIsTrue, claims, citations, aiSignals, annotations,
          community, category, checkedAt, processingSeconds,
        ]
      properties:
        id: { type: string }
        trackingId: { type: string }
        title: { type: string }
        contentType: { $ref: "#/components/schemas/ContentType" }
        language: { type: string }
        submittedText: { type: string }
        sourceUrl: { type: string, format: uri }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer, minimum: 0, maximum: 100, description: FR-DETECT-02. }
        summary: { type: string }
        whatIsFalse: { type: array, items: { type: string } }
        whatIsTrue: { type: array, items: { type: string } }
        claims: { type: array, items: { $ref: "#/components/schemas/Claim" } }
        citations: { type: array, items: { $ref: "#/components/schemas/Citation" } }
        aiSignals: { type: array, items: { $ref: "#/components/schemas/AISignal" } }
        annotations: { type: array, items: { $ref: "#/components/schemas/ExpertAnnotation" } }
        humanReview:
          oneOf:
            - { $ref: "#/components/schemas/HumanReview" }
            - type: "null"
        community: { $ref: "#/components/schemas/CommunityRating" }
        category: { type: string }
        checkedAt: { type: string, format: date-time }
        processingSeconds: { type: number }
...
    FactCheckPublic:
      description: |
        Partner-facing report shape (GET /v1/fact-checks/{id}). Matches the `factCheck()` helper
        in apps/web/components/developers/examples.ts exactly: no submittedText, no annotations,
        no rating comments — only aggregate community counts.
      type: object
      required:
        [
          id, trackingId, url, title, contentType, language, category, verdict, confidence,
          summary, whatIsFalse, whatIsTrue, claims, citations, aiSignals, community, humanReview,
          checkedAt, processingSeconds,
        ]
      properties:
        id: { type: string }
        trackingId: { type: string }
        url: { type: string, format: uri, example: "https://zuula.ug/fact-checks/fc-2026-0142" }
        title: { type: string }
        contentType: { $ref: "#/components/schemas/ContentType" }
        language: { type: string }
        category: { type: string }
        verdict: { $ref: "#/components/schemas/Verdict" }
        confidence: { type: integer }
        summary: { type: string }
        whatIsFalse: { type: array, items: { type: string } }
        whatIsTrue: { type: array, items: { type: string } }
        claims: { type: array, items: { $ref: "#/components/schemas/Claim" } }
        citations: { type: array, items: { $ref: "#/components/schemas/Citation" } }
        aiSignals: { type: array, items: { $ref: "#/components/schemas/AISignal" } }
        community:
          type: object
          required: [accurate, inaccurate]
          properties:
            accurate: { $ref: "#/components/schemas/RatingCounts" }
            inaccurate: { $ref: "#/components/schemas/RatingCounts" }
        humanReview:
          oneOf:
            - { $ref: "#/components/schemas/HumanReview" }
            - type: "null"
        checkedAt: { type: string, format: date-time }
        processingSeconds: { type: number }
...
    SubmissionInput:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [text, url, media, article] }
        content: { type: string, description: "Text (20–5,000 chars) or article body (50–20,000 chars)." }
        url: { type: string, format: uri }
        file: { type: string, format: binary, description: "Up to 50 MB; image/audio/video." }
        headline: { type: string, maxLength: 200 }
        articleUrl: { type: string, format: uri }
        language:
          allOf: [{ type: string }]
          description: "A LocaleCode, or `auto` to detect."
        captchaToken: { type: string, description: "Cloudflare Turnstile token, required when signed out (FR-AUTH-07)." }
    SubmissionAccepted:
      type: object
      required: [trackingId, status, estimatedSeconds, statusUrl]
      properties:
        trackingId: { type: string, example: ZL-7K3P-Q9 }
        status: { type: string, enum: [queued] }
        estimatedSeconds: { type: integer, example: 10 }
        statusUrl: { type: string, format: uri }
    SubmissionStepEvent:
      type: object
      required: [step, status]
      properties:
        step:
          type: string
          enum: [received, scan, fetch, transcribe, media, language, claims, sources, ai, report]
        status: { type: string, enum: [active, done] }
        seconds: { type: number, description: Expected duration for this step. }
    SubmissionStatusResponse:
      type: object
      required: [trackingId, status, submittedAt]
      properties:
        trackingId: { type: string }
        status: { type: string, enum: [queued, processing, completed, failed] }
        submittedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        steps:
          type: array
          items: { $ref: "#/components/schemas/SubmissionStepEvent" }
        result: { $ref: "#/components/schemas/FactCheckReport" }
        error: { $ref: "#/components/schemas/ErrorEnvelope" }
    ActivitySubmission:
      type: object
      required: [trackingId, type, preview, submittedAt, status]
      properties:
        trackingId: { type: string }
        type: { type: string, enum: [text, url, media, article] }
        preview: { type: string }
        submittedAt: { type: string, format: date-time }
        status: { type: string, enum: [processing, complete, failed] }
        verdict: { $ref: "#/components/schemas/Verdict" }
        reportId: { type: string }
...
    PartnerCheckInput:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [text, url, media, article] }
        content: { type: string }
        url: { type: string, format: uri }
        file: { type: string, format: binary }
        headline: { type: string }
        language: { type: string, description: "A LocaleCode, or `auto`." }
    PartnerCheckStatus:
      type: object
      required: [trackingId, status, submittedAt]
      properties:
        trackingId: { type: string }
        status: { type: string, enum: [queued, processing, completed, failed] }
        submittedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        result: { $ref: "#/components/schemas/FactCheckPublic" }
        error: { $ref: "#/components/schemas/ErrorEnvelope" }